refactor: Extract FunctionVisitor from c_parser.py
Extract FunctionVisitor class to separate module:
- Move ~375 lines of AST traversal logic to function_visitor.py
- Keep constants (C_KEYWORDS, AUTOSAR_TYPES, AUTOSAR_MACROS) on FunctionVisitor
- Re-export constants from CParser for backward compatibility
- c_parser.py reduced from ~1190 to ~745 lines
